+++ obj/src/connect/SqlServerConnectionResolver.js
@@ -1,5 +1,6 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
+exports.SqlServerConnectionResolver = void 0;
 /** @module connect */
 /** @hidden */
 const async = require('async');
@@ -16,19 +17,19 @@ const pip_services3_components_node_2 = require("pip-services3-components-node")
  *  ### Configuration parameters ###
  *
  * - connection(s):
- *   - discovery_key:               (optional) a key to retrieve the connection from [[https://rawgit.com/pip-services-node/pip-services3-components-node/master/doc/api/interfaces/connect.idiscovery.html IDiscovery]]
+ *   - discovery_key:               (optional) a key to retrieve the connection from [[https://pip-services3-node.github.io/pip-services3-components-node/interfaces/connect.idiscovery.html IDiscovery]]
  *   - host:                        host name or IP address
  *   - port:                        port number (default: 27017)
  *   - database:                    database name
  *   - uri:                         resource URI or connection string with all parameters in it
  * - credential(s):
- *   - store_key:                   (optional) a key to retrieve the credentials from [[https://rawgit.com/pip-services-node/pip-services3-components-node/master/doc/api/interfaces/auth.icredentialstore.html ICredentialStore]]
+ *   - store_key:                   (optional) a key to retrieve the credentials from [[https://pip-services3-node.github.io/pip-services3-components-node/interfaces/auth.icredentialstore.html ICredentialStore]]
  *   - username:                    user name
  *   - password:                    user password
  *
  * ### References ###
  *
- * - <code>\*:discovery:\*:\*:1.0</code>             (optional) [[https://rawgit.com/pip-services-node/pip-services3-components-node/master/doc/api/interfaces/connect.idiscovery.html IDiscovery]] services
+ * - <code>\*:discovery:\*:\*:1.0</code>             (optional) [[https://pip-services3-node.github.io/pip-services3-components-node/interfaces/connect.idiscovery.html IDiscovery]] services
  * - <code>\*:credential-store:\*:\*:1.0</code>      (optional) Credential stores to resolve credentials
  */
 class SqlServerConnectionResolver {