+++ obj/src/connect/index.js
@@ -1,9 +1,10 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
+exports.SqlServerConnectionResolver = void 0;
 /**
  * @module connect
  * @preferred
  */
 var SqlServerConnectionResolver_1 = require("./SqlServerConnectionResolver");
-exports.SqlServerConnectionResolver = SqlServerConnectionResolver_1.SqlServerConnectionResolver;
+Object.defineProperty(exports, "SqlServerConnectionResolver", { enumerable: true, get: function () { return SqlServerConnectionResolver_1.SqlServerConnectionResolver; } });
 //# sourceMappingURL=index.js.map