+++ obj/src/persistence/index.js
@@ -1,15 +1,16 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
+exports.IdentifiableJsonSqlServerPersistence = exports.IdentifiableSqlServerPersistence = exports.SqlServerPersistence = exports.SqlServerConnection = void 0;
 /**
  * @module persistence
  * @preferred
  */
 var SqlServerConnection_1 = require("./SqlServerConnection");
-exports.SqlServerConnection = SqlServerConnection_1.SqlServerConnection;
+Object.defineProperty(exports, "SqlServerConnection", { enumerable: true, get: function () { return SqlServerConnection_1.SqlServerConnection; } });
 var SqlServerPersistence_1 = require("./SqlServerPersistence");
-exports.SqlServerPersistence = SqlServerPersistence_1.SqlServerPersistence;
+Object.defineProperty(exports, "SqlServerPersistence", { enumerable: true, get: function () { return SqlServerPersistence_1.SqlServerPersistence; } });
 var IdentifiableSqlServerPersistence_1 = require("./IdentifiableSqlServerPersistence");
-exports.IdentifiableSqlServerPersistence = IdentifiableSqlServerPersistence_1.IdentifiableSqlServerPersistence;
+Object.defineProperty(exports, "IdentifiableSqlServerPersistence", { enumerable: true, get: function () { return IdentifiableSqlServerPersistence_1.IdentifiableSqlServerPersistence; } });
 var IdentifiableJsonSqlServerPersistence_1 = require("./IdentifiableJsonSqlServerPersistence");
-exports.IdentifiableJsonSqlServerPersistence = IdentifiableJsonSqlServerPersistence_1.IdentifiableJsonSqlServerPersistence;
+Object.defineProperty(exports, "IdentifiableJsonSqlServerPersistence", { enumerable: true, get: function () { return IdentifiableJsonSqlServerPersistence_1.IdentifiableJsonSqlServerPersistence; } });
 //# sourceMappingURL=index.js.map