+++ src/persistence/SqlServerPersistence.ts
@@ -108,7 +108,7 @@ export class SqlServerPersistence<T> implements IReferenceable, IUnreferenceable
     private _references: IReferences;
     private _opened: boolean;
     private _localConnection: boolean;
-    private _autoObjects: string[] = [];
+    private _schemaStatements: string[] = [];
 
     /**
      * The dependency resolver.
@@ -242,11 +242,34 @@ export class SqlServerPersistence<T> implements IReferenceable, IUnreferenceable
     }
 
     /**
-     * Adds index definition to create it on opening
-     * @param dmlStatement DML statement to autocreate database object
+     * Adds a statement to schema definition.
+     * This is a deprecated method. Use ensureSchema instead.
+     * @param schemaStatement a statement to be added to the schema
      */
-    protected autoCreateObject(dmlStatement: string): void {
-        this._autoObjects.push(dmlStatement);
+    protected autoCreateObject(schemaStatement: string): void {
+        this.ensureSchema(schemaStatement);
+    }
+
+    /**
+     * Adds a statement to schema definition
+     * @param schemaStatement a statement to be added to the schema
+     */
+    protected ensureSchema(schemaStatement: string): void {
+        this._schemaStatements.push(schemaStatement);
+    }
+
+    /**
+     * Clears all auto-created objects
+     */
+    protected clearSchema(): void {
+        this._schemaStatements = [];
+    }
+
+    /**
+     * Defines database schema via auto create objects or convenience methods.
+     */
+    protected defineSchema(): void {
+        // Todo: override in chile classes
     }
 
     /** 
@@ -321,8 +344,11 @@ export class SqlServerPersistence<T> implements IReferenceable, IUnreferenceable
                 this._databaseName = this._connection.getDatabaseName();
                 this._requestFactory = require('mssql').Request;
                 
+                 // Define database schema
+                 this.defineSchema();
+               
                 // Recreate objects
-                this.autoCreateObjects(correlationId, (err) => {
+                this.createSchema(correlationId, (err) => {
                     if (err) {
                         this._client == null;
                         err = new ConnectionException(correlationId, "CONNECT_FAILED", "Connection to sqlserver failed").withCause(err);    
@@ -401,8 +427,8 @@ export class SqlServerPersistence<T> implements IReferenceable, IUnreferenceable
         });
     }
 
-    protected autoCreateObjects(correlationId: string, callback: (err: any) => void): void {
-        if (this._autoObjects == null || this._autoObjects.length == 0) {
+    protected createSchema(correlationId: string, callback: (err: any) => void): void {
+        if (this._schemaStatements == null || this._schemaStatements.length == 0) {
             callback(null);
             return null;
         }
@@ -424,7 +450,7 @@ export class SqlServerPersistence<T> implements IReferenceable, IUnreferenceable
             this._logger.debug(correlationId, 'Table ' + this._tableName + ' does not exist. Creating database objects...');
 
             // Run all DML commands
-            async.eachSeries(this._autoObjects, (dml, callback) => {
+            async.eachSeries(this._schemaStatements, (dml, callback) => {
                 this._client.query(dml, (err, result) => {
                     if (err) {
                         this._logger.error(correlationId, err, 'Failed to autocreate database object');