+++ test/persistence/DummySqlServerPersistence.ts
@@ -12,8 +12,12 @@ export class DummySqlServerPersistence
 {
     public constructor() {
         super('dummies');
-        this.autoCreateObject('CREATE TABLE [dummies] ([id] VARCHAR(32) PRIMARY KEY, [key] VARCHAR(50), [content] VARCHAR(MAX))');
-        this.ensureIndex('dummies_key', { key: 1 }, { unique: true });
+    }
+
+    protected defineSchema(): void {
+        this.clearSchema();
+        this.ensureSchema('CREATE TABLE [' + this._tableName + '] ([id] VARCHAR(32) PRIMARY KEY, [key] VARCHAR(50), [content] VARCHAR(MAX))');
+        this.ensureIndex(this._tableName + '_key', { key: 1 }, { unique: true });
     }
 
     public getPageByFilter(correlationId: string, filter: FilterParams, paging: PagingParams,